animelist table and update type on watching status

+++ backend/db_scripts/create-table-watching_status.sql
@@ -1,5 +1,5 @@
 -- Active: 1715313977474@@localhost@5432@anilytics@public
 CREATE TABLE watching_status(  
-    status INT PRIMARY KEY,
+    status SMALLINT PRIMARY KEY,
     description VARCHAR(20) NOT NULL
 );
